// types/ethers-contracts/Coordinator.ts
export declare namespace BLS12381 {
  export type G2PointStruct = {
    word0: BytesLike;
    word1: BytesLike;
    word2: BytesLike;
  };

  export type G2PointStructOutput = [string, string, string] & {
    word0: string;
    word1: string;
    word2: string;
  };

  export type G1PointStruct = { word0: BytesLike; word1: BytesLike };

  export type G1PointStructOutput = [string, string] & {
    word0: string;
    word1: string;
  };
}

export declare namespace Coordinator {
  export type ParticipantStruct = {
    provider: string;
    aggregated: boolean;
    transcript: BytesLike;
    decryptionRequestStaticKey: BytesLike;
  };

  export type ParticipantStructOutput = [string, boolean, string, string] & {
    provider: string;
    aggregated: boolean;
    transcript: string;
    decryptionRequestStaticKey: string;
  };
}